requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.11.0
attrs==25.3.0
beautifulsoup4==4.14.2
bs4==0.0.2
certifi==2025.10.5
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
cssselect==1.3.0
datasets==4.1.1
dill==0.4.0
distro==1.9.0
docstring_parser==0.17.0
fastuuid==0.13.5
filelock==3.19.1
fire==0.7.1
frozenlist==1.7.0
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
litellm==1.77.7
lxml==6.0.2
lxml_html_clean==0.4.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.6.4
multiprocess==0.70.16
narwhals==2.6.0
nltk==3.9.2
numpy==2.3.3
openai==2.1.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
plotly==6.3.1
propcache==0.4.0
pyarrow==21.0.0
pydantic==2.11.10
pydantic_core==2.33.2
Pygments==2.19.2
PyMuPDF==1.26.4
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
readability==0.3.2
readability-lxml==0.8.1
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
selenium==4.27.1
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
termcolor==3.1.0
tiktoken==0.11.0
tokenizers==0.22.1
tqdm==4.67.1
trio==0.31.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wheel==0.45.1
wsproto==1.2.0
xxhash==3.6.0
yarl==1.21.0
zipp==3.23.0